Moved event utility functions out of tasks. Updated tests.
